Properly split smolvla preproc tests
